Better clipboard tool autodetection for linux/freebsd.
Old implementation relied on absence of `$DISPLAY` to detect Wayland.
But `$DISPLAY` can be set even in Wayland session. I don't fully know
why, I suspect it is because of XWayland.

So it won't pick correct tool.

This implementation relies on `$WAYLAND_DISPLAY`, which won't be set in
Xorg session, and `$XDG_SESSION_TYPE` to detect wayland. It also checks
if wl-clipboard is installed for consistency.

Additionally, it handles cases when there is no GUI running (SSH
session, or TTY).